requirements.txt
# numpy==1.24.3
numpy>=1.24.0,<2.0.0
# pandas==2.0.3
pandas>=2.0.0,<3.0.0
# scikit-learn==1.3.0
scikit-learn>=1.3.0,<2.0.0
# xgboost==2.0.0
xgboost>=2.0.0,<3.0.0
# lightgbm==4.1.0
lightgbm>=4.0.0,<5.0.0

# Big Data & Performance
dask[complete]>=2023.9.0
pyarrow>=13.0.0
fastparquet>=2023.7.0
joblib>=1.3.0

# Hyperparameter Optimization
optuna>=3.3.0

# Model Interpretability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0
folium>=0.14.0

# Geospatial
geopy>=2.4.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Streamlit App
streamlit>=1.28.0

# Development (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0
